refactor(native): isolate packed bit store selection
