+++ crates/celox-backend-arm64/src/isel.rs
@@ -4,6 +4,7 @@
 //! Supports 2-state and 4-state (IEEE 1800) with full mask propagation.
 //! Handles arbitrary widths: narrow (≤64-bit) and wide (>64-bit, chunk-based).
 
+mod packed_bit_store;
 mod sparse;
 
 use super::mir::*;
@@ -25,6 +26,7 @@ use celox_sir::analysis::{
     instruction_definition as sir_def_reg, reverse_postorder as ordered_sir_blocks,
     visit_instruction_uses as collect_sir_inst_uses,
 };
+use packed_bit_store::{PackedBitStorePlan, PackedBitStorePlans, find_packed_bit_store_plans};
 use sparse::{find_sparse_worklist_run, sparse_descriptor_table};
 
 /// Maps SIR RegisterId → MIR VReg for the current execution unit.
@@ -187,119 +189,6 @@ fn block_dynamic_load_cache_plans(
         })
         .collect();
     BlockDynamicLoadCachePlans { addresses }
-}
-
-#[derive(Debug, Clone)]
-struct PackedBitStorePlan {
-    source: RegisterId,
-    address: RegionedAbsoluteAddr,
-    first_lane: usize,
-    lane_count: usize,
-}
-
-#[derive(Debug, Default)]
-struct PackedBitStorePlans {
-    roots: HashMap<usize, PackedBitStorePlan>,
-    skip_indices: HashSet<usize>,
-}
-
-fn find_packed_bit_store_plans(
-    block: &BasicBlock<RegionedAbsoluteAddr>,
-    register_types: &HashMap<RegisterId, RegisterType>,
-    layout: &MemoryLayout,
-) -> PackedBitStorePlans {
-    let mut plans = PackedBitStorePlans::default();
-    let mut index = 0usize;
-    while index + 1 < block.instructions.len() {
-        let SIRInstruction::Slice(_, source, first_lane, 1) = block.instructions[index] else {
-            index += 1;
-            continue;
-        };
-        let SIRInstruction::Store(
-            address,
-            SIROffset::Static(first_store_lane),
-            1,
-            first_slice,
-            ref triggers,
-            ref captures,
-        ) = block.instructions[index + 1]
-        else {
-            index += 1;
-            continue;
-        };
-        let SIRInstruction::Slice(first_slice_definition, _, _, _) = block.instructions[index]
-        else {
-            unreachable!();
-        };
-        let Some(array) = layout.unpacked_arrays.get(&address.absolute_addr()) else {
-            index += 1;
-            continue;
-        };
-        if first_slice != first_slice_definition
-            || first_store_lane != first_lane
-            || !triggers.is_empty()
-            || !captures.is_empty()
-            || array.element_width != 1
-            || array.element_stride != 1
-        {
-            index += 1;
-            continue;
-        }
-
-        let mut lane_count = 0usize;
-        while index + lane_count * 2 + 1 < block.instructions.len() {
-            let slice_index = index + lane_count * 2;
-            let store_index = slice_index + 1;
-            let SIRInstruction::Slice(slice, lane_source, lane, 1) =
-                block.instructions[slice_index]
-            else {
-                break;
-            };
-            let SIRInstruction::Store(
-                lane_address,
-                SIROffset::Static(store_lane),
-                1,
-                stored,
-                ref lane_triggers,
-                ref lane_captures,
-            ) = block.instructions[store_index]
-            else {
-                break;
-            };
-            if lane_source != source
-                || lane != first_lane + lane_count
-                || lane_address != address
-                || store_lane != lane
-                || stored != slice
-                || !lane_triggers.is_empty()
-                || !lane_captures.is_empty()
-            {
-                break;
-            }
-            lane_count += 1;
-        }
-        let source_width = register_types.get(&source).map(RegisterType::width);
-        if lane_count >= 8
-            && lane_count.is_multiple_of(8)
-            && lane_count <= 64
-            && first_lane.is_multiple_of(8)
-            && source_width.is_some_and(|width| first_lane + lane_count <= width)
-            && first_lane + lane_count <= array.element_count
-        {
-            let plan = PackedBitStorePlan {
-                source,
-                address,
-                first_lane,
-                lane_count,
-            };
-            plans.skip_indices.extend(index..index + lane_count * 2);
-            plans.roots.insert(index, plan);
-            index += lane_count * 2;
-        } else {
-            index += 1;
-        }
-    }
-    plans
 }
 
 #[derive(Clone, Copy)]
